requirements.txt
asgiref==3.5.2; python_version >= "3.8"
attrs==22.1.0; python_version >= "3.7"
backports.zoneinfo==0.2.1; python_version < "3.9" and python_version >= "3.8"
decouple==0.0.7; python_version >= "3.7"
django-cors-headers==3.13.0; python_version >= "3.7"
django==4.1.1; python_version >= "3.8"
djangorestframework-simplejwt==5.2.0; python_version >= "3.7"
djangorestframework==3.14.0; python_version >= "3.6"
drf-spectacular==0.24.2; python_version >= "3.6"
importlib-resources==5.10.0; python_version < "3.9" and python_version >= "3.7"
inflection==0.5.1; python_version >= "3.6"
jsonschema==4.16.0; python_version >= "3.7"
pillow==9.2.0; python_version >= "3.7"
pkgutil-resolve-name==1.3.10; python_version < "3.9" and python_version >= "3.7"
pyjwt==2.5.0; python_version >= "3.7"
pyrsistent==0.18.1; python_version >= "3.7"
python-magic-bin==0.4.14
pytz==2022.2.1; python_version >= "3.7"
pyyaml==6.0; python_version >= "3.6"
sqlparse==0.4.3; python_version >= "3.8"
tzdata==2022.4; sys_platform == "win32" and python_version >= "3.8"
uritemplate==4.1.1; python_version >= "3.6"
zipp==3.9.0; python_version < "3.9" and python_version >= "3.7"
